+++ NiceComponentsExample/NiceComponentsExample/View/AllComponentsView.swift
@@ -33,7 +33,7 @@ struct AllComponentsView: View {
                     PrimaryButton("Primary Button") { }
                     SecondaryButton("Secondary Button") { }
                     BorderlessButton("Borderless Button") { }
-                    InactiveButton("Inactive Button") { }
+                    InactiveButton("Inactive Button")
                     DestructiveButton("Destructive Button") { }
                 }