+++ Sources/NiceComponents/Button/InactiveButton.swift
@@ -10,21 +10,20 @@ import SwiftUI
 public struct InactiveButton: View {
     let text: String
     let style: ButtonStyle
-    let onClick: () -> Void
 
-    public init(_ text: String, style: ButtonStyle? = nil, onClick: @escaping () -> Void) {
+    public init(_ text: String, style: ButtonStyle? = nil) {
         self.text = text
         self.style = style ?? Config.current.inactiveButtonStyle
-        self.onClick = onClick
     }
 
     public var body: some View {
-        Button(action: onClick) {
+        Button(action: {}) {
             Text(text)
                 .foregroundColor(style.onSurfaceColor)
                 .scaledFont(name: style.textStyle.name, size: style.textStyle.size, weight: style.textStyle.weight)
                 .frame(maxWidth: .infinity, minHeight: style.height, maxHeight: style.height)
         }
+        .disabled(true)
         .fixedSize(horizontal: false, vertical: true)
         .background(style.surfaceColor)
         .cornerRadius(style.border.radius)