+++ Sources/NiceComponents/Text/DetailText.swift
@@ -7,13 +7,17 @@
 
 import SwiftUI
 
-public struct DetailText: View {
+public struct DetailText: NiceText {
     public let text: String
     public let style: TypeStyle
 
+    public static var defaultStyle: TypeStyle {
+        Config.current.detailTextStyle
+    }
+
     public init(_ text: String, style: TypeStyle? = nil) {
         self.text = text
-        self.style = style ?? Config.current.detailTextStyle
+        self.style = style ?? Self.defaultStyle
     }
 
     public var body: some View {