+++ Sources/NiceComponents/Theme/Config.swift
@@ -49,6 +49,9 @@ public struct Config {
     public var screenTitleStyle: TypeStyle
     public var sectionTitleStyle: TypeStyle
 
+    /// Default is: x:0, y:4, blur: 4px, opacity: 0.15 (black)
+    public var shadowStyle: ShadowStyle
+
     public init(colorTheme: ColorTheme? = nil, typeTheme: TypeTheme? = nil) {
         self.colorTheme = colorTheme ?? ColorTheme()
         self.typeTheme = typeTheme ?? TypeTheme()
@@ -113,6 +116,15 @@ public struct Config {
         sectionTitleStyle = TypeStyle(
             color: self.colorTheme.onSurface,
             theme: self.typeTheme.headline2
+        )
+
+        // Set Shadow style
+
+        shadowStyle = ShadowStyle(
+            color: self.colorTheme.shadow,
+            radius: 4.0,
+            x: 0,
+            y: 4
         )
     }