+++ Sources/NiceComponents/Theme/Layout.swift
@@ -9,10 +9,17 @@ import SwiftUI
 
 public enum Layout {
     public enum Spacing {
+        /// xLarge: 64
         public static let xLarge: CGFloat = 64
+        /// large: 32
         public static let large: CGFloat = 32
+        /// medium: 24
+        public static let medium: CGFloat = 24
+        /// standard: 16
         public static let standard: CGFloat = 16
+        /// small: 8
         public static let small: CGFloat = 8
+        /// xSmall: 4
         public static let xSmall: CGFloat = 4
     }
 }