+++ Sources/NiceComponents/Title/ScreenTitle.swift
@@ -7,13 +7,18 @@
 
 import SwiftUI
 
-public struct ScreenTitle: View {
+public struct ScreenTitle: NiceText {
+
     public let text: String
     public let style: TypeStyle
 
+    public static var defaultStyle: TypeStyle {
+        Config.current.screenTitleStyle
+    }
+
     public init(_ text: String, style: TypeStyle? = nil) {
         self.text = text
-        self.style = style ?? Config.current.screenTitleStyle
+        self.style = style ?? Self.defaultStyle
     }
 
     public var body: some View {