+++ CHANGELOG.md
@@ -6,3 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 - Initial release! Adds a basic set of components and a first pass at config options to customize them.
+
+## [0.2.0]
+- Added shadowStyle 
+- Text components implement NiceText
+- Layout documentation
+- ContentLoadState is now equatable
+- ResizableImage handles both bundle string and URL
+- InactiveButton removed onClick modifier